Integrates task consolidate-normalization (0a12afa): drops the redundant double-normalization surfaced by the L3 mutation review (circuit_ir CircuitBuilder.gate .upper(), source/process pi1>=1.0 guard + dead coords init, source/coupling dtype, carrier/channels), re-syncing the D11/D12/D13/D15 batches. All 4 affected gates PASS in the merged state.

+++ src/error_coupling_simulator/source/coupling.py
@@ -303,7 +303,7 @@ def parameter_series(params: Iterable[CoupledMechanismParams], field: str) -> np
         if not hasattr(p, field):
             raise AttributeError(f"CoupledMechanismParams has no field {field!r}")
         values.append(float(getattr(p, field)))
-    return np.asarray(values, dtype=np.float64)
+    return np.asarray(values)
 
 
 def cross_mechanism_correlation(
